refactor(tenant-api): OIDC provider internal refactoring
