+++ packages/engine-tenant-api/src/model/service/idp/index.ts
@@ -4,4 +4,4 @@ export * from './IDPHandlerRegistry'
 export * from './IDPResponseError'
 export * from './IDPValidationError'
 export * from './InvalidIDPConfigurationError'
-export * from './OIDCProvider'
+export * from './providers/OIDCProvider'
